Move monorail specific issue tracker files to issue_management/monorail. (#511)
