+++ src/appengine/handlers/cron/oss_fuzz_build_status.py
@@ -24,7 +24,7 @@ from datastore import data_types
 from datastore import ndb
 from handlers import base_handler
 from issue_management import issue_tracker_utils
-from issue_management.issue import Issue
+from issue_management.monorail.issue import Issue
 from libs import handler
 from libs import helpers
 from metrics import logs
