+++ src/python/issue_management/issue_filer.py
@@ -18,7 +18,7 @@ from base import utils
 from datastore import data_handler
 from datastore import data_types
 from issue_management import label_utils
-from issue_management.issue import Issue
+from issue_management.monorail.issue import Issue
 from system import environment
 
 NON_CRASH_TYPES = [
