+++ src/python/tests/appengine/handlers/cron/oss_fuzz_build_status_test.py
@@ -24,7 +24,7 @@ import webtest
 
 from datastore import data_types
 from handlers.cron import oss_fuzz_build_status
-from issue_management.issue import Issue
+from issue_management.monorail.issue import Issue
 from tests.test_libs import helpers as test_helpers
 from tests.test_libs import test_utils
